Add caching, use $*TMPDIR
We add a simple caching layer so that we don't
hit the FS with mkdir and statcalls unnecessarily.

Also we wrap the sub-body in a $lock.protect call
so that we can avoid any weird accidents. In practice
this lock should hardly be much stressed unless
you have a distribution with dozens or hundreds of
resources that you are trying to load in all at once
on many threads.

That's the theory anyway.