Now it should actually work!
The %?RESOURCES of the consuming module need to be made
accessible.

The ways I thought of implementing this were as a export parameter

   use Resource::Wrangler resources => { $?RESOURCES };

or as a role parameter

   my $wrangler = Resource::Wrangler[{ %?RESOURCES }].new;

I opted for the latter for the sake of implementation
simplicity, but the former has some real appeal as well.

Who knows, maybe there will ve 3.0.0 with that change
soon :)